Header and  Side bar wrapped
