+++ front-end/src/components/Header.css
@@ -3,11 +3,22 @@
   background-color: black;
   color: white;
   display: flex;
-  /* height: 130px; */
 }
 
 .headTitleContainer {
   display: flex;
   flex-grow: 4;
   justify-content: center;
+}
+
+.sidebarwrapper {
+  display: flex;
+}
+.sidebar {
+  height: 100%;
+  width: 20%;
+}
+
+.maincomponent {
+  padding: 5px;
 }
