+++ front-end/src/components/SideBar.jsx
@@ -1,24 +1,37 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
+
 import './SideBar.css';
 
 const SideBar = () => {
   return (
-    <div className="sidebarContainer">
-<ul>
-  <li>
-          Produtos
-  </li>
-  <li>
-          Meu Pedidos
-  </li>
-  <li>
-          Meu Perfil
-  </li>
-  
-</ul>
-<h2>Sair</h2>
-</div>
-  
+    <div className="side-menu-container">
+      <ul>
+        <li>
+          <Link to={"/products"} data-testid="side-menu-item-products">
+            Produtos
+          </Link>
+        </li>
+        <li>
+          <Link to={"/orders"} data-testid="side-menu-item-my-orders">
+            Meu Pedidos
+          </Link>
+
+        </li>
+        <li>
+          <Link to={"/profile"} data-testid="side-menu-item-my-profile">
+            Meu Perfil
+          </Link>
+
+
+        </li>
+
+      </ul>
+      <Link to={"/"} data-testid="side-menu-item-logout">
+        <h2>Sair</h2>
+      </Link>
+    </div>
+
   )
 }
 export default SideBar;
