+++ front-end/src/pages/Products/Products.js
@@ -1,33 +1,11 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
-
 import Header from '../../components/Header';
-import SideBar from '../../components/SideBar';
 import ProductsContainer from '../../components/ProductsContainer';
 import './style.css';
 
 const Products = () => {
-  const sideBarVisible = useSelector(
-    (state) => state.sideBarHideReducer.isVisible,
-  );
-const MyComponent = ProductsContainer;
-
-
   return (
-    <>
-      {/* <Header wrapper={ProductsContainer}/> */}
-      <Header />
-
-      <div className="sidebarwrapper">
-        <div className="sidebar">{sideBarVisible && <SideBar />}</div>
-        <div className="maincomponent">
-          <h1> Here is Product.js</h1>
-          {/* <MyComponent /> */}
-
-          <ProductsContainer />
-        </div>
-      </div>
-    </>
+      <Header wrapper={ProductsContainer} /> //wrapper = component name to render
   );
 };
