--- front-end/src/pages/Products/style.css
@@ -1,17 +1,0 @@
-.sidebarwrapper {
-  display: flex;
-
-
-
-  
-}
-.sidebar{
-  /* background-color: black; */
-  width: 20%;
-  height: 100%;
-
-}
-
-.maincomponent{
-  padding: 5px;
-}
